feat: split contract into three different examples
