+++ Cargo.toml
@@ -1,16 +1,5 @@
-[package]
-name = "status-message"
-version = "0.1.0"
-authors = ["Near Inc <hello@nearprotocol.com>"]
-edition = "2018"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[dependencies]
-serde = { version = "*", features = ["derive"] }
-serde_json = "*"
-near-sdk = "3.1.0"
+[workspace]
+members = ["contracts/*"]
 
 [profile.release]
 codegen-units = 1
