+++ contracts/intermediate/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "intermediate"
+version = "0.0.1"
+authors = ["Near Inc <hello@nearprotocol.com>"]
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+near-sdk = "4.0.0"
+witgen = "0.12.0"
