+++ contracts/simple/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "simple"
+version = "0.0.1"
+authors = ["Near Inc <hello@nearprotocol.com>"]
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+near-sdk = "4.0.0"
